% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePaddedLabel.R
\name{create_padded_labels}
\alias{create_padded_labels}
\title{Create a vector of labels}
\usage{
create_padded_labels(n = 6, string = "Toy", suffix = NULL)
}
\arguments{
\item{n}{Number of labels to create.}

\item{string}{Character string for label.}

\item{suffix}{Optional character string suffix for file type.}
}
\value{
Vector of character strings with consecutive, padded labels.
    These labels could be names for vectors or file names.
    Padding of numeric values with leading zeroes ensures that
    file names will display in correct consecutive order in a file
    explorer (e.g. "Species09.csv" before "Species10.csv")
}
\description{
Create a vector of character strings with padded, consecutive labels.
}
\examples{
create_padded_labels()
create_padded_labels(n = 4,string = "Species")
create_padded_labels(n = 7,string= "Site" ,suffix = ".csv")
}
