% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTableTemplate.R
\name{data_table_template}
\alias{data_table_template}
\title{Title}
\usage{
data_table_template(data_frame = NULL, file_name = NULL)
}
\arguments{
\item{data_frame}{Name of data frame for template <keep as null>}

\item{file_name}{Name of the csv data file filename.csv}
}
\value{
named .csv
}
\description{
Title
}
\examples{
data_table_template()

}
