% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoLogConsole.R
\name{echo_log_console}
\alias{echo_log_console}
\title{Toggle Log File Messages to Echo to Console}
\usage{
echo_log_console(toggle = FALSE)
}
\arguments{
\item{toggle}{Logical variable setting whether log messages are echoed
in the console (\code{TRUE}) or sent only to the log file (\code{FALSE}).
Default is \code{FALSE}. Can be toggled at any point in a script.}
}
\value{
Toggles the switch.
}
\description{
Toggle Log File Messages to Echo to Console
}
\examples{
\dontrun{
echo_log_console(TRUE)
}

}
