% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitiateSeed.R
\name{initiate_seed}
\alias{initiate_seed}
\title{Automatically Select, Set, and Store Random Number Seed}
\usage{
initiate_seed(user_seed = NULL)
}
\arguments{
\item{user_seed}{Takes as input a random number seed from the user,
stores it, and sets the random number seed with the \code{set.seed()}
function. In the default state, the user does not provide a seed.
Instead, the function randomly selects an integer in the range +- 1e+9,
stores it, and invokes the \code{set.seed()}. This seed is recorded
in the log file each time the script is run. Automating the choice
of the random number seed avoids the problem that most users
select a seed that is very small (typically < 1e+3) compared
to the range of possible seed values that could be used (+- 1e+9).}
}
\value{
A randomly chosen integer (+- 1e+9) that
is internally stored, recorded, and passed
to the \code{set.seed()} function.
}
\description{
This function is used internally by the set_up_log function.
To set a seed manually, pass it to the set_up_log function with
set_up_log(user_seed=UserInteger)
}
\examples{
\dontrun{
initiate_seed()
initiate_seed(99)
}

}
