% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L.R
\name{l}
\alias{l}
\title{Pass Messages to Log File}
\usage{
l(log_entry = "")
}
\arguments{
\item{log_entry}{Text message to pass to log file. The default is
no message, in which case only the time stamp is entered in the log.}
}
\value{
Message sent to log.
}
\description{
Pass Messages to Log File
}
\examples{
\dontrun{
l('Text message for log')
}

}
