% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDataTemplate.R
\name{metadata_template}
\alias{metadata_template}
\title{Metadata Template For Embedded Annotation of Data Files}
\usage{
metadata_template(file_name = NULL)
}
\arguments{
\item{file_name}{File name for creation of .csv annotated data file.
If no name is provided, default is \emph{"MetaDataTemplate.csv"}.}
}
\value{
Creates date-stamped template file for data entry.
}
\description{
Metadata Template For Embedded Annotation of Data Files
}
\examples{
\dontrun{
metadata_template("HeffalumpData.csv")
}

}
