% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowProgressBar.R
\name{show_progress_bar}
\alias{show_progress_bar}
\title{Show Progress Bar}
\usage{
show_progress_bar(index = i, counter = 10, dot = 1)
}
\arguments{
\item{index}{integer variable used for counter within a loop
(typically \emph{i}, \emph{j}, or \emph{k}). The default value is \emph{i}.}

\item{counter}{Numeric label for the interval. This value will increment
a counter number that is displayed in the console.
The default value is 10,
 so the console will display 10,20,30,... when these repititions are
 reached in the for loop.}

\item{dot}{Increment value for each dot. The default value is 1, so a dot
is displayed for each pass through the for loop, with integers printed
after 9 consecutive dots (for the default value of counter=10).}
}
\value{
While a for loop is running, displays progress bar on the console, which is echoed to the log file.
}
\description{
Show Progress Bar
}
\examples{
\dontrun{
for (i in 1:100) {
   <for loop commands>
   show_progress_bar()
   }

for (j in 1:10000) {
   <for loop commands>
   show_progress_bar(index=j, counter=500, dot=50)
   }
 }
}
