% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceBatch.R
\name{source_batch}
\alias{source_batch}
\title{Batch Sourcing of Multiple R script files}
\usage{
source_batch(folder = NULL, filetype = "*.R$")
}
\arguments{
\item{folder}{Name of subfolder that contains R scripts to be sourced.
This folder name could be expanded to a path statement to source script
files in other locations. Scripts are sourced in alphabetical order
of their names. The function is not recursive so scripts contained within
subfolders are not sourced. The function fails if no folder is given,
if the folder does not exist, or if it is empty.}

\item{filetype}{Pattern to specify file type. Default is \emph{"*.R$"}.
To source all of the files in a folder as R scripts,
even if they do not contain the \emph{".R"} suffix, use \emph{"*.*$"}}
}
\value{
Runs the \code{source()} function on each R script file
contained in the folder.
}
\description{
Batch Sourcing of Multiple R script files
}
\examples{
\dontrun{
source_batch("Functions")
}

}
