% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetUpLog.R
\name{set_up_log}
\alias{set_up_log}
\title{Set Up Log File}
\usage{
set_up_log(my_logfile = "logfile.txt", console_echo = FALSE)
}
\arguments{
\item{my_logfile}{Character string for log file name.
Default is \emph{logfile.txt}.}

\item{console_echo}{Simple toggle}
}
\value{
Creates log file.
}
\description{
Set Up Log File
}
\examples{
\dontrun{
set_up_log()
}

}
